requirements.txt
streamlit
pdfplumber
easyocr
Pillow
python-docx
